prevent deletion of used ingredient

+++ backend/service/ingredient-service.go
@@ -1,19 +1,22 @@
 package service
 
 import (
+	"github.com/RemiEven/miam/common"
 	"github.com/RemiEven/miam/datasource"
 	"github.com/RemiEven/miam/model"
 )
 
 // IngredientService struct
 type IngredientService struct {
-	ingredientDao *datasource.IngredientDao
+	ingredientDao       *datasource.IngredientDao
+	recipeIngredientDao *datasource.RecipeIngredientDao
 }
 
 // newIngredientService creates a new ingredient service
-func newIngredientService(ingredientDao *datasource.IngredientDao) *IngredientService {
+func newIngredientService(ingredientDao *datasource.IngredientDao, recipeIngredientDao *datasource.RecipeIngredientDao) *IngredientService {
 	return &IngredientService{
 		ingredientDao,
+		recipeIngredientDao,
 	}
 }
 
@@ -36,5 +39,12 @@ func (service *IngredientService) UpdateIngredient(ID string, update model.BaseI
 
 // DeleteIngredient deletes the ingredient with the given id
 func (service *IngredientService) DeleteIngredient(ID string) error {
+	used, err := service.recipeIngredientDao.IsUsedInRecipe(ID)
+	if err != nil {
+		return err
+	}
+	if used {
+		return common.ErrInvalidOperation
+	}
 	return service.ingredientDao.DeleteIngredient(ID)
 }
